alx_travel_app/alx_travel_app/requirement.txt
mysqlclient==2.2.7
Django
djangorestframework
django-cors-headers
drf-yasg
mysqlclient
django-environ
celery
